pyproject.toml
[project]
name = "hica"
version = "0.1.0"
description = "A generalized agent library following 12-factor principles"
authors = [{ name = "Sandipan Haldar", email = "work@sandipanhaldar.com" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "instructor>=1.0",
    "openai>=1.0",
    "fastapi>=0.100",
    "uvicorn>=0.20",
    "python-dotenv>=1.0",
    "structlog>=24.0",
    "pydantic-ai>=0.2.19",
    "pytest>=8.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.20",
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "hica"
version = "0.1.0"
description = "A generalized agent library following 12-factor principles"
authors = ["Your Name <your.email@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.0"
instructor = "^1.0"
openai = "^1.0"
